# hospital-backend/swagger.yml
openapi: 3.0.0
info:
  title: Hospital Backend API
  description: API for a single hospital node in a federated learning network for breast cancer detection.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /hospital/info:
    get:
      tags: [Hospital]
      summary: Get hospital information
      description: Retrieve detailed information about this hospital node.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalInfoResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /hospital/stats:
    get:
      tags: [Hospital]
      summary: Get hospital statistics
      description: Retrieve patient and system statistics for this hospital.
      responses:
        '200':
          description: Successful response
        '500':
          $ref: '#/components/responses/InternalError'
...
  /blockchain/register-hospital:
    post:
      tags: [Blockchain]
      summary: Register hospital
      description: Register this hospital node on the blockchain smart contract.
      responses:
        '200':
          description: Hospital registered
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    HospitalInfo:
      type: object
      properties:
        hospital_id:
          type: string
        hospital_name:
          type: string
        capabilities:
          type: object
          properties:
            modalities:
              type: array
              items:
                type: string
            encryption_enabled:
              type: boolean
            ipfs_enabled:
              type: boolean
            blockchain_enabled:
              type: boolean
    HospitalInfoResponse:
      type: object
      properties:
        success:
          type: boolean
        hospital:
          $ref: '#/components/schemas/HospitalInfo'
...
  responses:
...
    InternalError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string